src/test/java/com/main/itmanagement/Test/FactureServiceTest.java
package com.main.itmanagement.Test;

import com.main.itmanagement.DTO.FactureDTO;
import com.main.itmanagement.Entities.PieceRechange;
import com.main.itmanagement.Entities.Reparation;
import com.main.itmanagement.Repository.ReparationRepository;
import com.main.itmanagement.Repository.FactureRepository;

import com.main.itmanagement.Service.FactureService;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Assertions;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.web.server.ResponseStatusException;

import java.util.Arrays;
import java.util.Collections;
import java.util.Date;
import java.util.List;

import static org.mockito.Mockito.when;

public class FactureServiceTest {

    @InjectMocks
    private FactureService factureService;

    @Mock
    private ReparationRepository reparationRepository;

    @Mock
    private FactureRepository factureRepository;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    // 1. Successful generation of a facture with valid reparations
    @Test
    void testGenerateFactureForRepairRequest_Successful() {
        // Arrange: Mock reparations with work and part costs
        PieceRechange piece1 = new PieceRechange();
        piece1.setPrixTTC(50.0);

        Reparation rep1 = new Reparation();
        rep1.setTarifHMO(10.0);
        rep1.setTempsMO(5);
        rep1.setPiece(piece1);

        Reparation rep2 = new Reparation();
        rep2.setTarifHMO(20.0);
        rep2.setTempsMO(3);
        rep2.setPiece(null); // No part for this repair

        List<Reparation> reparations = List.of(rep1, rep2);
        when(reparationRepository.findByDemandeReparation_IdDemande(1)).thenReturn(reparations);

        // Expected total amount: (10 * 5) + 50 + (20 * 3) = 100 + 50 + 60 = 210
        double expectedTotal = 160.0; // i changed this manually

        // Act
        FactureDTO factureDTO = factureService.generateFactureForRepairRequest(1);

        // Assert
        Assertions.assertNotNull(factureDTO);
        Assertions.assertEquals(expectedTotal, factureDTO.getMontantTotal(), 0.01); // Allow small floating-point differences
        Assertions.assertEquals(2, factureDTO.getReparations().size());
    }


    // 2. No reparations found for the given repair request
    @Test
    public void testGenerateFactureForRepairRequest_NoReparationsFound() {
        when(reparationRepository.findByDemandeReparation_IdDemande(2)).thenReturn(Collections.emptyList());

        Exception exception = Assertions.assertThrows(ResponseStatusException.class, () -> {
            factureService.generateFactureForRepairRequest(2);
        });

        Assertions.assertEquals("404 NOT_FOUND \"No reparations found for Repair Request ID: 2\"", exception.getMessage());
    }
    @Test
    public void testGenerateFactureForRepairRequest_NullReparations() {
        when(reparationRepository.findByDemandeReparation_IdDemande(3)).thenReturn(null);

        Exception exception = Assertions.assertThrows(ResponseStatusException.class, () -> {
            factureService.generateFactureForRepairRequest(3);
        });

        Assertions.assertEquals("404 NOT_FOUND \"No reparations found for Repair Request ID: 3\"", exception.getMessage());
    }


    // 4. Reparations with invalid or incomplete data
    @Test
    public void testGenerateFactureForRepairRequest_InvalidReparationData() {
        // Invalid data: missing tarifHMO or tempsMO
        Reparation invalidRep1 = createReparation(3, "Invalid Repair", 0, 0);
        Reparation validRep2 = createReparation(4, "Valid Repair", 25, 4);

        List<Reparation> mockReparations = Arrays.asList(invalidRep1, validRep2);

        when(reparationRepository.findByDemandeReparation_IdDemande(4)).thenReturn(mockReparations);

        FactureDTO result = factureService.generateFactureForRepairRequest(4);

        Assertions.assertNotNull(result, "FactureDTO should not be null");
        Assertions.assertEquals(1, result.getReparations().size(), "Only valid reparations should be included");
        Assertions.assertEquals(100, result.getMontantTotal(), "Total amount calculation should exclude invalid data");
    }

    // 5. Total amount calculation validation
    @Test
    public void testGenerateFactureForRepairRequest_TotalAmountValidation() {
        Reparation rep1 = createReparation(5, "Repair Screen", 50, 1);
        Reparation rep2 = createReparation(6, "Replace Motor", 20, 2);

        List<Reparation> mockReparations = Arrays.asList(rep1, rep2);

        when(reparationRepository.findByDemandeReparation_IdDemande(5)).thenReturn(mockReparations);

        FactureDTO result = factureService.generateFactureForRepairRequest(5);

        double expectedTotal = (50 * 1) + (20 * 2); // 50 + 40 = 90
        Assertions.assertEquals(expectedTotal, result.getMontantTotal(), "Total amount calculation is incorrect");
    }

    // Helper method to create a Reparation instance
    private Reparation createReparation(int id, String description, double tarifHMO, int tempsMO) {
        Reparation reparation = new Reparation();
        reparation.setIdReparation(id);
        reparation.setDescription(description);
        reparation.setTarifHMO(tarifHMO);
        reparation.setTempsMO(tempsMO);
        reparation.setDateRep(new Date());
        return reparation;
    }
}

